% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Implementation of the Dijkstra's algorithm
to find the shortest path between nodes
Wikipedia: https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{Graph to be traversed}

\item{init_node}{Initial node to start from}
}
\value{
numeric vector of shortest distances to nodes in given graph
}
\description{
Implementation of the Dijkstra's algorithm
to find the shortest path between nodes
Wikipedia: https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm
}
