% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclid.R
\name{euclid}
\alias{euclid}
\title{Implementation of the Euclidean algorithm
to find the greatest common divisor of two numbers}
\usage{
euclid(x, y)
}
\arguments{
\item{x}{numeric scalar or integer}

\item{y}{numeric scalar or integer}
}
\value{
greatest common divisor of x and y
}
\description{
Implementation of the Euclidean algorithm
to find the greatest common divisor of two numbers
}
\references{
https://en.wikipedia.org/wiki/Euclidean_algorithm
}
