% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{validate_input_dijkstra}
\alias{validate_input_dijkstra}
\title{Validate inputs for the dijkstra function}
\usage{
validate_input_dijkstra(graph, init_node)
}
\arguments{
\item{graph}{Graph to be traversed}

\item{init_node}{Initial node to start from}
}
\value{
error message in case input is not valid, empty string otherwise
}
\description{
Validate inputs for the dijkstra function
}
