% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_graph.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Example graph in edge list format (\strong{Wikipedia Dijkstra example})}
\format{
A data frame with 18 rows and 3 variables:
\describe{
\item{v1}{Node label (source node) — integer}
\item{v2}{Node label (target node) — integer}
\item{w}{Edge weight (non-negative numeric)}
}
}
\source{
\url{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}
}
\usage{
wiki_graph
}
\description{
Example graph in edge list format (\strong{Wikipedia Dijkstra example})
}
\keyword{datasets}
